SP 204: Add bypass tray support
The printer is annoying for this one.  Unless you actually print to the bypass
tray, it will ignore it, so this must be set in the printer options.  The tray
is also annoying to use in that it's a manual feed, accepting one sheet at a
time.

Anyway, add the option to specify the bypass tray in the PPD and filter.

Signed-off-by: James Bottomley <JBottomley@Parallels.com>